+++ DarkCodex/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.0.2]
+
+### Added
+- fix hexcrafter lv3 hex arcana selection; lv11 spell recall
+- cold iron arrows refresh after combat
+- a kineticist background
+- healing burn heals non-lethal damage (instead of turning it into real damage)
+- kinetic blasts apply gather power, even if a weaker mode is picked
+- mobile gathering
+- made gather power ability visible
+
 ## [1.0.1]
 
 ### Added